fix: package simulation runtime for integration
