+++ projects/policyengine-api-simulation/src/modal/app.py
@@ -2,7 +2,7 @@
 PolicyEngine Simulation - Versioned Modal App
 
 This app contains the heavy simulation workload with snapshotted models.
-Each deployment creates a versioned app (e.g., policyengine-simulation-us1-459-0-uk2-65-9).
+Each deployment creates a versioned app (e.g., policyengine-simulation-py4-10-0).
 
 The gateway app (policyengine-simulation-gateway) routes requests to these versioned apps.
 """
@@ -13,7 +13,7 @@ import os
 from src.modal._image_setup import snapshot_models
 from src.modal.dependency_pins import project_dependency_pin
 from src.modal.logging_redaction import redact_params_for_logging
-from src.modal.release_bundle import get_bundled_country_model_version
+from policyengine_api_simulation.release_bundle import get_bundled_country_model_version
 
 POLICYENGINE_VERSION = os.environ.get("POLICYENGINE_VERSION") or project_dependency_pin(
     "policyengine"
@@ -31,20 +31,19 @@ UK_VERSION = os.environ.get(
 ) or get_bundled_country_model_version("uk")
 
 
-def get_app_name(us_version: str, uk_version: str) -> str:
+def get_app_name(policyengine_version: str) -> str:
     """
-    Generate versioned app name from package versions.
+    Generate versioned app name from the policyengine.py package version.
 
     Replaces dots with dashes for URL safety.
-    Example: us1.459.0, uk2.65.9 -> policyengine-simulation-us1-459-0-uk2-65-9
+    Example: 4.10.0 -> policyengine-simulation-py4-10-0
     """
-    us_safe = us_version.replace(".", "-")
-    uk_safe = uk_version.replace(".", "-")
-    return f"policyengine-simulation-us{us_safe}-uk{uk_safe}"
+    policyengine_safe = policyengine_version.replace(".", "-")
+    return f"policyengine-simulation-py{policyengine_safe}"
 
 
 # App name can be overridden via environment variable, otherwise generated from versions
-APP_NAME = os.environ.get("MODAL_APP_NAME", get_app_name(US_VERSION, UK_VERSION))
+APP_NAME = os.environ.get("MODAL_APP_NAME", get_app_name(POLICYENGINE_VERSION))
 
 # App definition with versioned name
 app = modal.App(APP_NAME)
@@ -67,7 +66,11 @@ simulation_image = (
         "tables>=3.10.2",
         "logfire",
     )
-    .add_local_python_source("src.modal", copy=True)
+    .add_local_python_source(
+        "src.modal",
+        "policyengine_api_simulation",
+        copy=True,
+    )
     .run_function(snapshot_models)
 )
 
@@ -106,7 +109,7 @@ def run_simulation(params: dict) -> dict:
     """
     import logfire
 
-    from src.modal.simulation import run_simulation_impl
+    from policyengine_api_simulation.simulation_runtime import run_simulation_impl
 
     configure_logfire()
