+++ projects/policyengine-api-simulation/src/modal/hf_dataset.py
@@ -1,0 +1,3 @@
+"""Compatibility shim for packaged simulation helpers."""
+
+from policyengine_api_simulation.hf_dataset import *  # noqa: F403
