+++ projects/policyengine-api-simulation/src/modal/utils/update_version_registry.py
@@ -1,8 +1,10 @@
 """
 Update Modal version registries after deployment.
 
-Each deployment creates a versioned app (e.g., policyengine-simulation-us1-459-0-uk2-65-9).
-This script updates the version dicts to map package versions to app names.
+Each deployment creates a versioned policyengine.py app (e.g.,
+policyengine-simulation-py4-10-0). This script updates the version dicts to map
+the policyengine.py version and the bundled country package versions to that
+app name.
 
 The dicts allow the gateway to route requests for specific versions to the correct app.
 Multiple versions can coexist - old deployments remain accessible via their version numbers.
@@ -12,7 +14,8 @@ to point to the new app. This ensures "latest" behavior uses the most recent dep
 
 Usage:
     uv run python -m src.modal.utils.update_version_registry \
-        --app-name policyengine-simulation-us1-459-0-uk2-65-9 \
+        --app-name policyengine-simulation-py4-10-0 \
+        --policyengine-version 4.10.0 \
         --us-version 1.459.0 \
         --uk-version 2.65.9 \
         --environment staging
@@ -21,6 +24,11 @@ Usage:
 import argparse
 import modal
 from packaging.version import InvalidVersion, Version
+
+POLICYENGINE_VERSION_DICT_NAME = "simulation-api-policyengine-versions"
+US_VERSION_DICT_NAME = "simulation-api-us-versions"
+UK_VERSION_DICT_NAME = "simulation-api-uk-versions"
+APP_RELEASE_BUNDLES_DICT_NAME = "simulation-api-app-release-bundles"
 
 
 def _is_newer_version(candidate: str, current: str | None) -> bool:
@@ -108,6 +116,59 @@ def update_version_dict(
         )
 
 
+def _country_bundle_metadata(country: str) -> dict:
+    from policyengine_api_simulation.release_bundle import (
+        DATASET_ALIASES,
+        get_country_release_bundle,
+    )
+
+    bundle = get_country_release_bundle(country)
+    return {
+        "country": bundle.country,
+        "model_package_name": bundle.model_package_name,
+        "model_version": bundle.model_version,
+        "data_package_name": bundle.data_package_name,
+        "data_version": bundle.data_version,
+        "default_dataset": bundle.default_dataset,
+        "default_dataset_uri": bundle.default_dataset_uri,
+        "dataset_uris": dict(bundle.dataset_uris),
+        "dataset_aliases": dict(DATASET_ALIASES.get(bundle.country, {})),
+    }
+
+
+def build_app_release_bundle_metadata(
+    *,
+    app_name: str,
+    policyengine_version: str,
+) -> dict:
+    return {
+        "app_name": app_name,
+        "policyengine_version": policyengine_version,
+        "us": _country_bundle_metadata("us"),
+        "uk": _country_bundle_metadata("uk"),
+    }
+
+
+def put_app_release_bundle_metadata(
+    *,
+    environment: str,
+    app_name: str,
+    policyengine_version: str,
+) -> None:
+    bundle_store = modal.Dict.from_name(
+        APP_RELEASE_BUNDLES_DICT_NAME,
+        environment_name=environment,
+        create_if_missing=True,
+    )
+    metadata = build_app_release_bundle_metadata(
+        app_name=app_name,
+        policyengine_version=policyengine_version,
+    )
+    bundle_store[app_name] = metadata
+    bundle_store[policyengine_version] = metadata
+    print(f"  {APP_RELEASE_BUNDLES_DICT_NAME}[{app_name}]: updated")
+
+
 def main():
     parser = argparse.ArgumentParser(
         description="Update version registries after Modal deployment"
@@ -115,7 +176,12 @@ def main():
     parser.add_argument(
         "--app-name",
         required=True,
-        help="Versioned app name (e.g., policyengine-simulation-us1-459-0-uk2-65-9)",
+        help="Versioned app name (e.g., policyengine-simulation-py4-10-0)",
+    )
+    parser.add_argument(
+        "--policyengine-version",
+        required=True,
+        help="policyengine.py package version (e.g., 4.10.0)",
     )
     parser.add_argument(
         "--us-version",
@@ -144,14 +210,25 @@ def main():
 
     print(f"Updating version registries in Modal environment: {args.environment}")
     print(f"  App name: {args.app_name}")
+    print(f"  policyengine.py version: {args.policyengine_version}")
     print(f"  US version: {args.us_version}")
     print(f"  UK version: {args.uk_version}")
+    print()
+
+    print("policyengine.py version registry:")
+    update_version_dict(
+        POLICYENGINE_VERSION_DICT_NAME,
+        args.environment,
+        args.policyengine_version,
+        args.app_name,
+        force_latest=args.force_latest,
+    )
     print()
 
     # Update US registry
     print("US version registry:")
     update_version_dict(
-        "simulation-api-us-versions",
+        US_VERSION_DICT_NAME,
         args.environment,
         args.us_version,
         args.app_name,
@@ -162,11 +239,19 @@ def main():
     # Update UK registry
     print("UK version registry:")
     update_version_dict(
-        "simulation-api-uk-versions",
+        UK_VERSION_DICT_NAME,
         args.environment,
         args.uk_version,
         args.app_name,
         force_latest=args.force_latest,
+    )
+    print()
+
+    print("App release bundle metadata:")
+    put_app_release_bundle_metadata(
+        environment=args.environment,
+        app_name=args.app_name,
+        policyengine_version=args.policyengine_version,
     )
     print()
