+++ projects/policyengine-api-simulation/src/policyengine_api_simulation/simulation_runtime.py
@@ -1,0 +1,513 @@
+"""
+Simulation implementation - pure logic with snapshotted imports.
+
+This module avoids importing policyengine at module level so the worker can
+load the requested country module without triggering cross-country imports.
+No Modal dependencies here.
+"""
+
+import contextlib
+import json
+import logging
+import os
+import tempfile
+from dataclasses import dataclass
+from importlib import import_module
+from typing import Any, Iterator
+
+from policyengine_api_simulation.hf_dataset import with_hf_revision
+from policyengine_api_simulation.release_bundle import (
+    get_country_release_bundle,
+    resolve_bundle_dataset_name,
+    resolve_bundle_dataset_uri,
+)
+from policyengine_api_simulation.simulation_output_builder import (
+    SimulationOutputBuilder,
+)
+from policyengine_api_simulation.telemetry import split_internal_payload
+
+logger = logging.getLogger(__name__)
+
+os.environ.setdefault("POLICYENGINE_SKIP_COUNTRY_IMPORTS", "1")
+
+DEFAULT_YEAR = 2026
+
+
+@dataclass(frozen=True)
+class RegionResolution:
+    code: str
+    dataset_reference: str | None = None
+    scoping_strategy: Any | None = None
+
+
+def _normalize_credentials_blob(creds_json: str) -> str:
+    """Return the raw JSON blob, decoding the outer escape if present.
+
+    The upstream Modal secret sometimes stores the credentials payload
+    double-encoded (the entire JSON object is wrapped in quotes with
+    backslash-escaped interior quotes). Historically we always attempted
+    the unescape as a fallback which could accidentally parse an already
+    clean blob. Only unwrap when the payload looks wrapped."""
+
+    try:
+        json.loads(creds_json)
+    except json.JSONDecodeError:
+        looks_escaped = creds_json.lstrip().startswith('"') or '\\"' in creds_json
+        if looks_escaped:
+            return json.loads(f'"{creds_json}"')
+        raise
+    return creds_json
+
+
+@contextlib.contextmanager
+def setup_gcp_credentials() -> Iterator[None]:
+    """
+    Set up GCP credentials from environment variable.
+
+    Modal secrets are injected as environment variables. The GCP library
+    expects GOOGLE_APPLICATION_CREDENTIALS to point to a file path. If
+    credentials JSON is provided, write it to a temp file that's deleted
+    on exit. This runs as a context manager to guarantee cleanup even if
+    the caller raises mid-simulation; the previous fire-and-forget
+    ``tempfile.mkstemp`` path leaked credential material on disk every
+    time a container served a request.
+    """
+    # Log available GCP-related env vars for debugging
+    gcp_vars = {
+        k: v[:50] + "..." if len(v) > 50 else v
+        for k, v in os.environ.items()
+        if "GOOGLE" in k or "GCP" in k or "CREDENTIAL" in k
+    }
+    logger.info(f"GCP-related env vars: {list(gcp_vars.keys())}")
+
+    # Check if credentials are already set as a file path
+    if os.environ.get("GOOGLE_APPLICATION_CREDENTIALS"):
+        logger.info("GOOGLE_APPLICATION_CREDENTIALS already set")
+        yield
+        return
+
+    # Check for credentials JSON in various env var names
+    creds_json = (
+        os.environ.get("GOOGLE_APPLICATION_CREDENTIALS_JSON")
+        or os.environ.get("GCP_CREDENTIALS_JSON")
+        or os.environ.get("GOOGLE_CREDENTIALS")
+        or os.environ.get("SERVICE_ACCOUNT_JSON")
+    )
+
+    if not creds_json:
+        logger.warning("No GCP credentials found in environment variables")
+        yield
+        return
+
+    normalized = _normalize_credentials_blob(creds_json)
+
+    # ``NamedTemporaryFile(delete=True)`` removes the file when the context
+    # exits (either normally or via exception). We restore any prior value
+    # of ``GOOGLE_APPLICATION_CREDENTIALS`` so a retry in the same
+    # container doesn't silently pick up a path that no longer exists.
+    previous = os.environ.get("GOOGLE_APPLICATION_CREDENTIALS")
+    with tempfile.NamedTemporaryFile(
+        mode="w", suffix=".json", delete=True
+    ) as creds_file:
+        creds_file.write(normalized)
+        creds_file.flush()
+        os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = creds_file.name
+        logger.info(f"GCP credentials written to {creds_file.name}")
+        try:
+            yield
+        finally:
+            if previous is None:
+                os.environ.pop("GOOGLE_APPLICATION_CREDENTIALS", None)
+            else:
+                os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = previous
+
+
+def run_simulation_impl(params: dict) -> dict:
+    """
+    Execute economic simulation.
+
+    Pure implementation with no Modal dependencies.
+    Accepts the gateway simulation payload and returns the legacy macro result dict.
+    """
+    # Set up GCP credentials if needed. The credentials temp file is
+    # cleaned up on exit so we never leave signed JSON material on disk.
+    with setup_gcp_credentials():
+        return _run_simulation_impl_core(params)
+
+
+def _parse_year(params: dict[str, Any]) -> int:
+    value = params.get("time_period") or params.get("year") or DEFAULT_YEAR
+    return int(value)
+
+
+def _normalise_period_key(period_key: Any) -> str:
+    """Convert legacy ``start.stop`` period keys to v4 effective dates."""
+    text = str(period_key)
+    parts = text.split(".")
+    if len(parts) > 1 and len(parts[0]) == 10:
+        return parts[0]
+    return text
+
+
+def _normalise_policy(policy: dict[str, Any] | None) -> dict[str, Any] | None:
+    if not policy:
+        return None
+
+    normalised: dict[str, Any] = {}
+    for parameter, value in policy.items():
+        if isinstance(value, dict):
+            normalised[parameter] = {
+                _normalise_period_key(period): period_value
+                for period, period_value in value.items()
+            }
+        else:
+            normalised[parameter] = value
+    return normalised
+
+
+def _resolve_dataset_name(country: str, requested_data: str | None) -> str:
+    return resolve_bundle_dataset_name(country, requested_data)
+
+
+def _split_requested_revision(requested_data: str) -> tuple[str, str | None]:
+    if "@" not in requested_data:
+        return requested_data, None
+    dataset_name, revision = requested_data.rsplit("@", maxsplit=1)
+    if not dataset_name or not revision:
+        raise ValueError(f"Invalid dataset revision reference: {requested_data}")
+    return dataset_name, revision
+
+
+def _requested_data_version(params: dict[str, Any]) -> str | None:
+    data_version = params.get("data_version")
+    if data_version is not None:
+        return str(data_version)
+
+    data = params.get("data")
+    if isinstance(data, str) and "@" in data:
+        _, revision = _split_requested_revision(data)
+        return revision
+    return None
+
+
+def _resolve_dataset_reference(country: str, params: dict[str, Any]) -> str:
+    requested_data = params.get("data")
+    requested_data = requested_data if isinstance(requested_data, str) else None
+    requested_data_version = _requested_data_version(params)
+
+    if requested_data_version is None:
+        return _resolve_dataset_name(country, requested_data)
+
+    if requested_data is None:
+        dataset_uri = get_country_release_bundle(country).default_dataset_uri
+    else:
+        dataset_without_revision, data_revision = _split_requested_revision(
+            requested_data
+        )
+        if data_revision is not None and data_revision != requested_data_version:
+            raise ValueError(
+                "Conflicting dataset revisions: "
+                f"data requests {data_revision!r} but data_version is "
+                f"{requested_data_version!r}"
+            )
+        dataset_uri = resolve_bundle_dataset_uri(country, dataset_without_revision)
+
+    return with_hf_revision(dataset_uri, requested_data_version)
+
+
+def _normalise_region_code(country: str, region: Any) -> str:
+    if region is None or str(region).strip() == "":
+        return country
+
+    raw = str(region).strip()
+    if raw.lower() in {"us", "uk"}:
+        return raw.lower()
+
+    if "/" not in raw:
+        if country == "us" and len(raw) == 2:
+            return f"state/{raw.lower()}"
+        if country == "uk":
+            return f"country/{raw.lower().replace(' ', '_')}"
+        return raw
+
+    prefix, value = raw.split("/", maxsplit=1)
+    prefix = prefix.lower()
+    value = value.strip()
+    if prefix == "state":
+        value = value.lower()
+    elif prefix == "country":
+        value = value.lower().replace(" ", "_")
+    elif prefix in {"congressional_district", "place"}:
+        value = value.upper()
+    elif prefix == "local_authority":
+        value = value.upper()
+    return f"{prefix}/{value}"
+
+
+def _build_uk_weight_replacement_region(region_code: str):
+    if "/" not in region_code:
+        return None
+
+    prefix, value = region_code.split("/", maxsplit=1)
+    if prefix not in {"constituency", "local_authority"}:
+        return None
+
+    from policyengine.core.region import Region
+    from policyengine.core.scoping_strategy import WeightReplacementStrategy
+    from policyengine.data.uk_geography_assets import (
+        CONSTITUENCY_ASSET_SPEC,
+        LOCAL_AUTHORITY_ASSET_SPEC,
+    )
+
+    asset_spec = (
+        CONSTITUENCY_ASSET_SPEC
+        if prefix == "constituency"
+        else LOCAL_AUTHORITY_ASSET_SPEC
+    )
+    return Region(
+        code=region_code,
+        label=value,
+        region_type=prefix,
+        parent_code="uk",
+        scoping_strategy=WeightReplacementStrategy(
+            weight_matrix_bucket=asset_spec.bucket,
+            weight_matrix_key=asset_spec.weight_matrix_filename,
+            lookup_csv_bucket=asset_spec.bucket,
+            lookup_csv_key=asset_spec.lookup_csv_filename,
+            region_code=value,
+        ),
+    )
+
+
+def _region_parent_dataset_reference(
+    country_module,
+    country: str,
+    region,
+    params: dict[str, Any],
+) -> str:
+    parent_code = getattr(region, "parent_code", None)
+    if parent_code:
+        parent_region = country_module.model.get_region(parent_code)
+        parent_dataset_path = getattr(parent_region, "dataset_path", None)
+        if isinstance(parent_dataset_path, str):
+            requested_data_version = _requested_data_version(params)
+            return (
+                with_hf_revision(parent_dataset_path, requested_data_version)
+                if requested_data_version is not None
+                else parent_dataset_path
+            )
+    return _resolve_dataset_reference(country, params)
+
+
+def _resolve_region(
+    *,
+    country_module,
+    country: str,
+    params: dict[str, Any],
+) -> RegionResolution:
+    region_code = _normalise_region_code(country, params.get("region"))
+    if region_code == country:
+        return RegionResolution(
+            code=region_code,
+            dataset_reference=_resolve_dataset_reference(country, params),
+        )
+
+    region = country_module.model.get_region(region_code)
+    if region is None and country == "uk":
+        region = _build_uk_weight_replacement_region(region_code)
+    if region is None:
+        raise ValueError(f"Unsupported {country.upper()} region: {region_code}")
+
+    dataset_path = getattr(region, "dataset_path", None)
+    requested_data_version = _requested_data_version(params)
+    if isinstance(dataset_path, str):
+        dataset_reference = (
+            with_hf_revision(dataset_path, requested_data_version)
+            if requested_data_version is not None
+            else dataset_path
+        )
+    else:
+        dataset_reference = _region_parent_dataset_reference(
+            country_module,
+            country,
+            region,
+            params,
+        )
+
+    return RegionResolution(
+        code=region_code,
+        dataset_reference=dataset_reference,
+        scoping_strategy=getattr(region, "scoping_strategy", None),
+    )
+
+
+def _microframe_like(frame, weights: str):
+    from microdf import MicroDataFrame
+
+    return MicroDataFrame(frame.copy(), weights=weights)
+
+
+def _person_group_column(person, entity: str) -> str:
+    prefixed = f"person_{entity}_id"
+    if prefixed in person.columns:
+        return prefixed
+    return f"{entity}_id"
+
+
+def _subsample_us_dataset(dataset, subsample: int | None):
+    if not subsample:
+        return dataset
+
+    from policyengine.tax_benefit_models.us.datasets import (
+        PolicyEngineUSDataset,
+        USYearData,
+    )
+
+    dataset.load()
+    data = dataset.data
+    household = data.household.head(int(subsample)).copy()
+    household_ids = set(household["household_id"])
+
+    person_household_col = _person_group_column(data.person, "household")
+    person = data.person[data.person[person_household_col].isin(household_ids)].copy()
+
+    def group_subset(entity: str):
+        person_col = _person_group_column(person, entity)
+        entity_id_col = f"{entity}_id"
+        ids = set(person[person_col])
+        frame = getattr(data, entity)
+        return frame[frame[entity_id_col].isin(ids)].copy()
+
+    subset_data = USYearData(
+        person=_microframe_like(person, "person_weight"),
+        marital_unit=_microframe_like(
+            group_subset("marital_unit"), "marital_unit_weight"
+        ),
+        family=_microframe_like(group_subset("family"), "family_weight"),
+        spm_unit=_microframe_like(group_subset("spm_unit"), "spm_unit_weight"),
+        tax_unit=_microframe_like(group_subset("tax_unit"), "tax_unit_weight"),
+        household=_microframe_like(household, "household_weight"),
+    )
+    subset_path = os.path.join(
+        os.environ.get("POLICYENGINE_DATA_FOLDER", "/tmp/policyengine-data"),
+        f"{dataset.id}_subsample_{subsample}.h5",
+    )
+    return PolicyEngineUSDataset(
+        id=f"{dataset.id}_subsample_{subsample}",
+        name=f"{dataset.name} subsample {subsample}",
+        description=dataset.description,
+        filepath=subset_path,
+        year=dataset.year,
+        is_output_dataset=dataset.is_output_dataset,
+        metadata=getattr(dataset, "metadata", {}),
+        metadata_filepath=getattr(dataset, "metadata_filepath", None),
+        data=subset_data,
+    )
+
+
+def _country_module(country: str):
+    country = country.lower()
+    if country not in {"us", "uk"}:
+        raise ValueError(f"Unsupported country: {country}")
+
+    return import_module(f"policyengine.tax_benefit_models.{country}")
+
+
+def _load_dataset(
+    params: dict[str, Any],
+    *,
+    country_module=None,
+    region_resolution: RegionResolution | None = None,
+):
+    country = params.get("country", "us").lower()
+    year = _parse_year(params)
+    country_module = country_module or _country_module(country)
+    dataset_name = (
+        region_resolution.dataset_reference
+        if region_resolution is not None and region_resolution.dataset_reference
+        else _resolve_dataset_reference(country, params)
+    )
+    datasets = country_module.ensure_datasets(
+        datasets=[dataset_name],
+        years=[year],
+        data_folder=os.environ.get(
+            "POLICYENGINE_DATA_FOLDER", "/tmp/policyengine-data"
+        ),
+    )
+    dataset = next(iter(datasets.values()))
+    if country == "us":
+        return _subsample_us_dataset(dataset, params.get("subsample"))
+    return dataset
+
+
+def _build_simulation(
+    params: dict[str, Any],
+    *,
+    dataset,
+    policy: dict[str, Any] | None,
+    scoping_strategy=None,
+):
+    from policyengine.core import Simulation
+
+    country_module = _country_module(params.get("country", "us"))
+    return Simulation(
+        dataset=dataset,
+        tax_benefit_model_version=country_module.model,
+        policy=policy,
+        scoping_strategy=scoping_strategy,
+    )
+
+
+def _run_simulation_impl_core(params: dict) -> dict:
+    simulation_params, telemetry, metadata = split_internal_payload(params)
+
+    logger.info(
+        "Starting simulation for country=%s run_id=%s process_id=%s",
+        simulation_params.get("country", "unknown"),
+        getattr(telemetry, "run_id", None),
+        getattr(telemetry, "process_id", None),
+    )
+    if metadata:
+        logger.info("Received simulation metadata keys: %s", sorted(metadata))
+
+    country = simulation_params.get("country", "us").lower()
+    country_module = _country_module(country)
+    region_resolution = _resolve_region(
+        country_module=country_module,
+        country=country,
+        params=simulation_params,
+    )
+    dataset = _load_dataset(
+        simulation_params,
+        country_module=country_module,
+        region_resolution=region_resolution,
+    )
+    baseline_policy = _normalise_policy(simulation_params.get("baseline"))
+    reform_policy = _normalise_policy(simulation_params.get("reform"))
+
+    logger.info("Initialising baseline and reform simulations")
+    baseline = _build_simulation(
+        simulation_params,
+        dataset=dataset,
+        policy=baseline_policy,
+        scoping_strategy=region_resolution.scoping_strategy,
+    )
+    reform = _build_simulation(
+        simulation_params,
+        dataset=dataset,
+        policy=reform_policy,
+        scoping_strategy=region_resolution.scoping_strategy,
+    )
+
+    logger.info("Calculating economic impact")
+    output = SimulationOutputBuilder(
+        country=country,
+        simulation_params=simulation_params,
+        country_module=country_module,
+        dataset=dataset,
+        baseline=baseline,
+        reform=reform,
+        resolved_data_version=_requested_data_version(simulation_params),
+    ).serialize()
+    logger.info("Comparison complete")
+    return output
