+++ projects/policyengine-api-simulation/tests/gateway/test_endpoints.py
@@ -8,7 +8,8 @@ simulation requests.
 import pytest
 from fastapi.testclient import TestClient
 
-from src.modal.release_bundle import resolve_bundle_dataset_uri
+from fixtures.gateway.test_endpoints import resolve_test_dataset_uri
+from policyengine_api_simulation.hf_dataset import HuggingFaceDatasetReferenceError
 
 
 def expected_bundle(
@@ -18,11 +19,18 @@ def expected_bundle(
     dataset: str | None = None,
     data_version: str | None = None,
 ) -> dict[str, str | None]:
+    resolved_dataset = resolve_test_dataset_uri(country, dataset)
+    if (
+        data_version is not None
+        and resolved_dataset is not None
+        and resolved_dataset.startswith("hf://")
+    ):
+        resolved_dataset = (
+            f"{resolved_dataset.rsplit('@', maxsplit=1)[0]}@{data_version}"
+        )
     bundle: dict[str, str | None] = {
         "model_version": model_version,
-        "dataset": resolve_bundle_dataset_uri(country, dataset)
-        if dataset is not None
-        else None,
+        "dataset": resolved_dataset,
     }
     if data_version is not None:
         bundle["data_version"] = data_version
@@ -43,7 +51,7 @@ class TestGetAppName:
         # Given
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         # When
@@ -51,7 +59,7 @@ class TestGetAppName:
 
         # Then
         assert resolved_version == "1.500.0"
-        assert app_name == "policyengine-simulation-us1-500-0-uk2-66-0"
+        assert app_name == "policyengine-simulation-py4-10-0"
 
     def test__given_us_country_with_version__then_returns_specified_app(
         self, mock_modal
@@ -65,7 +73,7 @@ class TestGetAppName:
 
         # Given
         mock_modal["dicts"]["simulation-api-us-versions"] = {
-            "1.459.0": "policyengine-simulation-us1-459-0-uk2-65-9"
+            "1.459.0": "policyengine-simulation-py3-9-0"
         }
 
         # When
@@ -73,7 +81,7 @@ class TestGetAppName:
 
         # Then
         assert resolved_version == "1.459.0"
-        assert app_name == "policyengine-simulation-us1-459-0-uk2-65-9"
+        assert app_name == "policyengine-simulation-py3-9-0"
 
     def test__given_uk_country__then_uses_uk_version_dict(self, mock_modal):
         """
@@ -86,7 +94,7 @@ class TestGetAppName:
         # Given
         mock_modal["dicts"]["simulation-api-uk-versions"] = {
             "latest": "2.66.0",
-            "2.66.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "2.66.0": "policyengine-simulation-py4-10-0",
         }
 
         # When
@@ -94,7 +102,7 @@ class TestGetAppName:
 
         # Then
         assert resolved_version == "2.66.0"
-        assert app_name == "policyengine-simulation-us1-500-0-uk2-66-0"
+        assert app_name == "policyengine-simulation-py4-10-0"
 
     def test__given_invalid_country__then_raises_value_error(self):
         """
@@ -138,7 +146,7 @@ class TestSubmitSimulationEndpoint:
         # Given
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -154,7 +162,7 @@ class TestSubmitSimulationEndpoint:
         # Then
         assert response.status_code == 200
         assert mock_modal["func"].last_from_name_call == (
-            "policyengine-simulation-us1-500-0-uk2-66-0",
+            "policyengine-simulation-py4-10-0",
             "run_simulation",
         )
 
@@ -169,7 +177,7 @@ class TestSubmitSimulationEndpoint:
         # Given
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -184,7 +192,7 @@ class TestSubmitSimulationEndpoint:
         # Then
         assert response.status_code == 200
         assert mock_modal["func"].last_from_name_call == (
-            "policyengine-simulation-us1-500-0-uk2-66-0",
+            "policyengine-simulation-py4-10-0",
             "run_simulation",
         )
         assert "time_period" not in mock_modal["func"].last_payload
@@ -201,7 +209,7 @@ class TestSubmitSimulationEndpoint:
         # Given
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -231,7 +239,7 @@ class TestSubmitSimulationEndpoint:
         """
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -263,7 +271,7 @@ class TestSubmitSimulationEndpoint:
         # Given
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -279,7 +287,7 @@ class TestSubmitSimulationEndpoint:
         # Then
         assert response.status_code == 200
         data = response.json()
-        assert data["resolved_app_name"] == "policyengine-simulation-us1-500-0-uk2-66-0"
+        assert data["resolved_app_name"] == "policyengine-simulation-py4-10-0"
         assert data["policyengine_bundle"] == expected_bundle(
             "us",
             "1.500.0",
@@ -291,7 +299,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -305,16 +313,91 @@ class TestSubmitSimulationEndpoint:
 
         assert response.status_code == 200
         data = response.json()
-        assert data["policyengine_bundle"]["dataset"] == resolve_bundle_dataset_uri(
+        assert data["policyengine_bundle"]["dataset"] == resolve_test_dataset_uri(
             "us", "enhanced_cps_2024"
         )
+
+    def test__given_submission_with_logical_revision__then_bundle_dataset_uses_revision(
+        self, mock_modal, client: TestClient
+    ):
+        mock_modal["dicts"]["simulation-api-us-versions"] = {
+            "latest": "1.500.0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
+        }
+
+        response = client.post(
+            "/simulate/economy/comparison",
+            json={
+                "country": "us",
+                "scope": "macro",
+                "reform": {},
+                "data": "enhanced_cps_2024@1.77.0",
+            },
+        )
+
+        assert response.status_code == 200
+        assert response.json()["policyengine_bundle"]["dataset"] == (
+            "hf://policyengine/policyengine-us-data/enhanced_cps_2024.h5@1.77.0"
+        )
+
+    def test__given_submission_with_conflicting_data_versions__then_returns_400(
+        self, mock_modal, client: TestClient
+    ):
+        mock_modal["dicts"]["simulation-api-us-versions"] = {
+            "latest": "1.500.0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
+        }
+
+        response = client.post(
+            "/simulate/economy/comparison",
+            json={
+                "country": "us",
+                "scope": "macro",
+                "reform": {},
+                "data": "enhanced_cps_2024@1.77.0",
+                "data_version": "1.78.2",
+            },
+        )
+
+        assert response.status_code == 400
+        assert mock_modal["func"].last_payload is None
+
+    def test__given_submission_with_invalid_hf_revision__then_returns_400_before_spawn(
+        self, mock_modal, client: TestClient, monkeypatch
+    ):
+        mock_modal["dicts"]["simulation-api-us-versions"] = {
+            "latest": "1.500.0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
+        }
+
+        def reject_revision(dataset_uri, revision):
+            raise HuggingFaceDatasetReferenceError("revision missing")
+
+        monkeypatch.setattr(
+            "src.modal.gateway.endpoints.with_hf_revision",
+            reject_revision,
+        )
+
+        response = client.post(
+            "/simulate/economy/comparison",
+            json={
+                "country": "us",
+                "scope": "macro",
+                "reform": {},
+                "data": "enhanced_cps_2024@does-not-exist",
+            },
+        )
+
+        assert response.status_code == 400
+        assert response.json()["detail"] == "revision missing"
+        assert mock_modal["func"].last_payload is None
 
     def test__given_submission_with_uk_alias_data__then_bundle_dataset_is_versioned_uri(
         self, mock_modal, client: TestClient
     ):
         mock_modal["dicts"]["simulation-api-uk-versions"] = {
             "latest": "2.66.0",
-            "2.66.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "2.66.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -328,7 +411,7 @@ class TestSubmitSimulationEndpoint:
 
         assert response.status_code == 200
         data = response.json()
-        assert data["policyengine_bundle"]["dataset"] == resolve_bundle_dataset_uri(
+        assert data["policyengine_bundle"]["dataset"] == resolve_test_dataset_uri(
             "uk", "enhanced_frs"
         )
 
@@ -337,7 +420,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -372,7 +455,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         request_body = {
@@ -399,7 +482,7 @@ class TestSubmitSimulationEndpoint:
         # Given
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
@@ -420,7 +503,7 @@ class TestSubmitSimulationEndpoint:
         data = response.json()
         assert data["status"] == "complete"
         assert "run_id" not in data
-        assert data["resolved_app_name"] == "policyengine-simulation-us1-500-0-uk2-66-0"
+        assert data["resolved_app_name"] == "policyengine-simulation-py4-10-0"
         assert data["policyengine_bundle"] == expected_bundle(
             "us",
             "1.500.0",
@@ -432,7 +515,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
@@ -490,7 +573,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
@@ -514,7 +597,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
@@ -542,7 +625,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
@@ -568,7 +651,7 @@ class TestSubmitSimulationEndpoint:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
@@ -600,7 +683,7 @@ class TestBudgetWindowBatchEndpoints:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         response = client.post(
@@ -618,7 +701,7 @@ class TestBudgetWindowBatchEndpoints:
 
         assert response.status_code == 200
         assert mock_modal["func"].last_from_name_call == (
-            "policyengine-simulation-us1-500-0-uk2-66-0",
+            "policyengine-simulation-py4-10-0",
             "run_budget_window_batch",
         )
         assert response.json() == {
@@ -627,7 +710,7 @@ class TestBudgetWindowBatchEndpoints:
             "poll_url": "/budget-window-jobs/mock-batch-job-id-123",
             "country": "us",
             "version": "1.500.0",
-            "resolved_app_name": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "resolved_app_name": "policyengine-simulation-py4-10-0",
             "policyengine_bundle": expected_bundle("us", "1.500.0"),
         }
 
@@ -636,7 +719,7 @@ class TestBudgetWindowBatchEndpoints:
     ):
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
@@ -672,7 +755,7 @@ class TestBudgetWindowBatchEndpoints:
             "child_jobs": {},
             "result": None,
             "error": None,
-            "resolved_app_name": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "resolved_app_name": "policyengine-simulation-py4-10-0",
             "policyengine_bundle": expected_bundle("us", "1.500.0"),
             "run_id": "batch-run-123",
         }
@@ -697,7 +780,7 @@ class TestBudgetWindowBatchEndpoints:
                 region="us",
                 version="1.500.0",
                 target="general",
-                resolved_app_name="policyengine-simulation-us1-500-0-uk2-66-0",
+                resolved_app_name="policyengine-simulation-py4-10-0",
                 policyengine_bundle=PolicyEngineBundle(model_version="1.500.0"),
                 start_year="2026",
                 window_size=2,
@@ -880,7 +963,7 @@ class TestBudgetWindowBatchEndpoints:
 
         mock_modal["dicts"]["simulation-api-us-versions"] = {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
 
         submit_response = client.post(
