+++ projects/policyengine-api-simulation/tests/gateway/test_package_imports.py
@@ -1,4 +1,5 @@
 import sys
+import importlib
 
 
 def test_gateway_models_import_does_not_import_fastapi_endpoints(
@@ -9,3 +10,22 @@ def test_gateway_models_import_does_not_import_fastapi_endpoints(
 
     assert gateway_import_module_names.endpoints not in sys.modules
     assert gateway_import_module_names.fastapi not in sys.modules
+
+
+def test_gateway_endpoints_import_does_not_import_policyengine_bundle(
+    isolated_gateway_model_import_modules,
+    gateway_import_module_names,
+):
+    release_bundle_module = sys.modules.pop(
+        "policyengine_api_simulation.release_bundle", None
+    )
+
+    try:
+        importlib.import_module(gateway_import_module_names.endpoints)
+
+        assert "policyengine_api_simulation.release_bundle" not in sys.modules
+    finally:
+        if release_bundle_module is not None:
+            sys.modules["policyengine_api_simulation.release_bundle"] = (
+                release_bundle_module
+            )
