+++ projects/policyengine-api-simulation/tests/test_modal_telemetry.py
@@ -1,4 +1,4 @@
-from src.modal.telemetry import TelemetryEnvelope, split_internal_payload
+from policyengine_api_simulation.telemetry import TelemetryEnvelope, split_internal_payload
 
 
 def test_split_internal_payload__removes_internal_fields():
