+++ projects/policyengine-api-simulation/tests/test_standalone_simulation_contract.py
@@ -1,9 +1,32 @@
 """Contract tests for the live synchronous simulation FastAPI app."""
 
+from importlib import import_module
+from pathlib import Path
+
 from fastapi.testclient import TestClient
 
 from fixtures.test_simulation_api_contracts import CURRENT_SINGLE_YEAR_MACRO_RESULT
 from policyengine_api_simulation.main import app
+
+
+PACKAGED_RUNTIME_MODULES = (
+    "policyengine_api_simulation.compat_models",
+    "policyengine_api_simulation.hf_dataset",
+    "policyengine_api_simulation.release_bundle",
+    "policyengine_api_simulation.simulation",
+    "policyengine_api_simulation.simulation_macro_output",
+    "policyengine_api_simulation.simulation_output_builder",
+    "policyengine_api_simulation.simulation_runtime",
+    "policyengine_api_simulation.telemetry",
+)
+
+
+def test_standalone_package_runtime_does_not_import_unpackaged_modal_source():
+    for module_name in PACKAGED_RUNTIME_MODULES:
+        module = import_module(module_name)
+        source = Path(module.__file__).read_text(encoding="utf-8")
+
+        assert "src.modal" not in source
 
 
 def test_standalone_simulation_openapi_keeps_legacy_schema_names():
