+++ projects/policyengine-api-simulation/tests/test_update_version_registry.py
@@ -64,7 +64,7 @@ def test_update_version_dict__keeps_latest_when_incoming_older(patched_modal):
     stores["main/simulation-api-us-versions"] = FakeDict(
         {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
     )
 
@@ -72,12 +72,12 @@ def test_update_version_dict__keeps_latest_when_incoming_older(patched_modal):
         "simulation-api-us-versions",
         "main",
         "1.400.0",
-        "policyengine-simulation-us1-400-0-uk2-66-0",
+        "policyengine-simulation-py3-8-0",
     )
 
     snapshot = stores["main/simulation-api-us-versions"].snapshot()
     assert snapshot["latest"] == "1.500.0"
-    assert snapshot["1.400.0"] == "policyengine-simulation-us1-400-0-uk2-66-0"
+    assert snapshot["1.400.0"] == "policyengine-simulation-py3-8-0"
 
 
 def test_update_version_dict__advances_latest_when_incoming_newer(patched_modal):
@@ -85,7 +85,7 @@ def test_update_version_dict__advances_latest_when_incoming_newer(patched_modal)
     stores["main/simulation-api-us-versions"] = FakeDict(
         {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
     )
 
@@ -93,7 +93,7 @@ def test_update_version_dict__advances_latest_when_incoming_newer(patched_modal)
         "simulation-api-us-versions",
         "main",
         "1.601.2",
-        "policyengine-simulation-us1-601-2-uk2-70-0",
+        "policyengine-simulation-py4-11-0",
     )
 
     snapshot = stores["main/simulation-api-us-versions"].snapshot()
@@ -105,7 +105,7 @@ def test_update_version_dict__force_latest_allows_downgrade(patched_modal):
     stores["main/simulation-api-us-versions"] = FakeDict(
         {
             "latest": "1.500.0",
-            "1.500.0": "policyengine-simulation-us1-500-0-uk2-66-0",
+            "1.500.0": "policyengine-simulation-py4-10-0",
         }
     )
 
@@ -113,7 +113,7 @@ def test_update_version_dict__force_latest_allows_downgrade(patched_modal):
         "simulation-api-us-versions",
         "main",
         "1.400.0",
-        "policyengine-simulation-us1-400-0-uk2-66-0",
+        "policyengine-simulation-py3-8-0",
         force_latest=True,
     )
 
@@ -128,9 +128,41 @@ def test_update_version_dict__new_registry_sets_latest_even_without_force(
         "simulation-api-uk-versions",
         "staging",
         "2.66.0",
-        "policyengine-simulation-us1-500-0-uk2-66-0",
+        "policyengine-simulation-py4-10-0",
     )
 
     snapshot = patched_modal["staging/simulation-api-uk-versions"].snapshot()
     assert snapshot["latest"] == "2.66.0"
-    assert snapshot["2.66.0"] == "policyengine-simulation-us1-500-0-uk2-66-0"
+    assert snapshot["2.66.0"] == "policyengine-simulation-py4-10-0"
+
+
+def test_put_app_release_bundle_metadata_records_app_and_py_version_aliases(
+    patched_modal,
+    monkeypatch,
+):
+    def fake_country_bundle_metadata(country: str) -> dict:
+        return {
+            "country": country,
+            "model_version": "1.0.0" if country == "us" else "2.0.0",
+            "data_version": "3.0.0" if country == "us" else "4.0.0",
+            "dataset_uris": {"default": f"hf://datasets/policyengine/{country}"},
+            "dataset_aliases": {"alias": "default"},
+        }
+
+    monkeypatch.setattr(
+        registry, "_country_bundle_metadata", fake_country_bundle_metadata
+    )
+
+    registry.put_app_release_bundle_metadata(
+        environment="main",
+        app_name="policyengine-simulation-py4-10-0",
+        policyengine_version="4.10.0",
+    )
+
+    snapshot = patched_modal[
+        "main/simulation-api-app-release-bundles"
+    ].snapshot()
+    metadata = snapshot["policyengine-simulation-py4-10-0"]
+    assert snapshot["4.10.0"] == metadata
+    assert metadata["policyengine_version"] == "4.10.0"
+    assert metadata["us"]["dataset_aliases"] == {"alias": "default"}
